Move utility modules into src/utils/
