+++ src/checkbox-field/checkbox-field.stories.mdx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { Meta, Story, Canvas, ArgsTable, Description } from '@storybook/addon-docs'
 import { BADGE } from '@geometricpanda/storybook-addon-badges'
 
-import { PartialProps } from '../storybook-helper'
+import { PartialProps } from '../utils/storybook-helper'
 import { CheckboxField } from '.'
 import { Stack } from '../stack'
 import { Text } from '../text'
