+++ src/hidden/hidden.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Box } from '../box'
 import styles from './hidden.module.css'
-import type { ResponsiveBreakpoints } from '../responsive-props'
+import type { ResponsiveBreakpoints } from '../utils/responsive-props'
 import { polymorphicComponent } from '../utils/polymorphism'
 
 type AboveProp = {
