+++ src/icons/alert-icon.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import type { AlertTone } from '../common-types'
+import type { AlertTone } from '../utils/common-types'
 
 const alertIconForTone: Record<AlertTone, typeof AlertInfoIcon> = {
     info: AlertInfoIcon,
