+++ src/inline/inline.stories.tsx
@@ -8,7 +8,7 @@ import {
     selectWithNone,
     PartialProps,
     disableResponsiveProps,
-} from '../storybook-helper'
+} from '../utils/storybook-helper'
 import { Stack } from '../stack'
 import { Heading } from '../heading'
 import { Inline, InlineAlign } from './inline'
