+++ src/inline/inline.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
 import { polymorphicComponent } from '../utils/polymorphism'
-import { mapResponsiveProp } from '../responsive-props'
+import { mapResponsiveProp } from '../utils/responsive-props'
 import { Box } from '../box'
 
-import type { ResponsiveProp } from '../responsive-props'
-import type { Space } from '../common-types'
+import type { ResponsiveProp } from '../utils/responsive-props'
+import type { Space } from '../utils/common-types'
 import type { ReusableBoxProps } from '../box'
 
 type InlineAlign = 'left' | 'center' | 'right'
