+++ src/stack/stack.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react'
 import flattenChildren from 'react-keyed-flatten-children'
 import { polymorphicComponent } from '../utils/polymorphism'
-import { mapResponsiveProp } from '../responsive-props'
+import { mapResponsiveProp } from '../utils/responsive-props'
 import { Box } from '../box'
 import { Divider } from '../divider'
 
-import type { ResponsiveProp } from '../responsive-props'
-import type { DividerWeight, Space } from '../common-types'
+import type { ResponsiveProp } from '../utils/responsive-props'
+import type { DividerWeight, Space } from '../utils/common-types'
 import type { BoxProps, ReusableBoxProps } from '../box'
 
 type Align = 'start' | 'center' | 'end'
