+++ src/text-link/text-link.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Box } from '../box'
 import { polymorphicComponent } from '../utils/polymorphism'
 import styles from './text-link.module.css'
-import type { OpenInNewTab } from '../common-types'
+import type { OpenInNewTab } from '../utils/common-types'
 
 type TextLinkProps = OpenInNewTab
