+++ src/utils/common-helpers.ts
